+++ changelog.md
@@ -4,6 +4,9 @@
 ## [Unrealeased]
 
 ### Added
+- !endgame ${lobby-id} will now delete that role for the server.
+- !endall will now delete all current roles given to player. 
+- Players will now be added to a game library when they start the game.
 - The game has awoken and has the sky ( the night at least ).
 - Have added functionality to start a lobby, and to join the lobby.
 - Also have added the functionality to roll the dice.